Test against Gradle nightly with scala3 support
