+++ core/src/main/scala/kafka/network/Acceptor.scala
@@ -1,0 +1,245 @@
+package kafka.network
+
+import java.nio.channels.{Selector as NSelector, _}
+import kafka.cluster.EndPoint
+import kafka.metrics.KafkaMetricsGroup
+import kafka.network.Processor.ListenerMetricTag
+import kafka.utils.CoreUtils
+import org.apache.kafka.common.KafkaException
+import org.apache.kafka.common.network.Selectable
+import org.apache.kafka.common.utils.{KafkaThread, Time, Utils}
+import org.slf4j.event.Level
+
+import java.net.{InetSocketAddress, SocketException}
+import java.nio.channels.{SelectionKey, ServerSocketChannel, SocketChannel}
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.atomic.AtomicBoolean
+import scala.collection.mutable.{ArrayBuffer, Buffer}
+import scala.collection.{Map, Seq, mutable}
+import scala.util.control.ControlThrowable
+
+
+/**
+ * Thread that accepts and configures new connections. There is one of these per endpoint.
+ */
+private[kafka] class Acceptor(val endPoint: EndPoint,
+                              val sendBufferSize: Int,
+                              val recvBufferSize: Int,
+                              nodeId: Int,
+                              connectionQuotas: ConnectionQuotas,
+                              metricPrefix: String,
+                              time: Time,
+                              logPrefix: String = "") extends AbstractServerThread(connectionQuotas) with KafkaMetricsGroup {
+
+  this.logIdent = logPrefix
+  private val nioSelector = NSelector.open()
+  val serverChannel = openServerSocket(endPoint.host, endPoint.port)
+  private val processors = new ArrayBuffer[Processor]()
+  private val processorsStarted = new AtomicBoolean
+  private val blockedPercentMeter = newMeter(s"${metricPrefix}AcceptorBlockedPercent",
+    "blocked time", TimeUnit.NANOSECONDS, Map(ListenerMetricTag -> endPoint.listenerName.value))
+  private var currentProcessorIndex = 0
+  private[network] val throttledSockets = new mutable.PriorityQueue[DelayedCloseSocket]()
+
+  private[network] case class DelayedCloseSocket(socket: SocketChannel, endThrottleTimeMs: Long) extends Ordered[DelayedCloseSocket] {
+    override def compare(that: DelayedCloseSocket): Int = endThrottleTimeMs compare that.endThrottleTimeMs
+  }
+
+  private[kafka] def addProcessors(newProcessors: Buffer[Processor], processorThreadPrefix: String): Unit = synchronized {
+    processors ++= newProcessors
+    if (processorsStarted.get)
+      startProcessors(newProcessors, processorThreadPrefix)
+  }
+
+  private[network] def startProcessors(processorThreadPrefix: String): Unit = synchronized {
+    if (!processorsStarted.getAndSet(true)) {
+      startProcessors(processors, processorThreadPrefix)
+    }
+  }
+
+  private def startProcessors(processors: Seq[Processor], processorThreadPrefix: String): Unit = synchronized {
+    processors.foreach { processor =>
+      KafkaThread.nonDaemon(
+        s"${processorThreadPrefix}-kafka-network-thread-$nodeId-${endPoint.listenerName}-${endPoint.securityProtocol}-${processor.id}",
+        processor
+      ).start()
+    }
+  }
+
+  private[network] def removeProcessors(removeCount: Int, requestChannel: RequestChannel): Unit = synchronized {
+    // Shutdown `removeCount` processors. Remove them from the processor list first so that no more
+    // connections are assigned. Shutdown the removed processors, closing the selector and its connections.
+    // The processors are then removed from `requestChannel` and any pending responses to these processors are dropped.
+    val toRemove = processors.takeRight(removeCount)
+    processors.remove(processors.size - removeCount, removeCount)
+    toRemove.foreach(_.initiateShutdown())
+    toRemove.foreach(_.awaitShutdown())
+    toRemove.foreach(processor => requestChannel.removeProcessor(processor.id))
+  }
+
+  override def initiateShutdown(): Unit = {
+    super.initiateShutdown()
+    synchronized {
+      processors.foreach(_.initiateShutdown())
+    }
+  }
+
+  override def awaitShutdown(): Unit = {
+    super.awaitShutdown()
+    synchronized {
+      processors.foreach(_.awaitShutdown())
+    }
+  }
+
+  /**
+   * Accept loop that checks for new connection attempts
+   */
+  def run(): Unit = {
+    serverChannel.register(nioSelector, SelectionKey.OP_ACCEPT)
+    startupComplete()
+    try {
+      while (isRunning) {
+        try {
+          acceptNewConnections()
+          closeThrottledConnections()
+        }
+        catch {
+          // We catch all the throwables to prevent the acceptor thread from exiting on exceptions due
+          // to a select operation on a specific channel or a bad request. We don't want
+          // the broker to stop responding to requests from other clients in these scenarios.
+          case e: ControlThrowable => throw e
+          case e: Throwable => error("Error occurred", e)
+        }
+      }
+    } finally {
+      debug("Closing server socket, selector, and any throttled sockets.")
+      CoreUtils.swallow(serverChannel.close(), this, Level.ERROR)
+      CoreUtils.swallow(nioSelector.close(), this, Level.ERROR)
+      throttledSockets.foreach(throttledSocket => closeSocket(throttledSocket.socket))
+      throttledSockets.clear()
+      shutdownComplete()
+    }
+  }
+
+  /**
+   * Create a server socket to listen for connections on.
+   */
+  private def openServerSocket(host: String, port: Int): ServerSocketChannel = {
+    val socketAddress =
+      if (Utils.isBlank(host))
+        new InetSocketAddress(port)
+      else
+        new InetSocketAddress(host, port)
+    val serverChannel = ServerSocketChannel.open()
+    serverChannel.configureBlocking(false)
+    if (recvBufferSize != Selectable.USE_DEFAULT_BUFFER_SIZE)
+      serverChannel.socket().setReceiveBufferSize(recvBufferSize)
+
+    try {
+      serverChannel.socket.bind(socketAddress)
+      info(s"Awaiting socket connections on ${socketAddress.getHostString}:${serverChannel.socket.getLocalPort}.")
+    } catch {
+      case e: SocketException =>
+        throw new KafkaException(s"Socket server failed to bind to ${socketAddress.getHostString}:$port: ${e.getMessage}.", e)
+    }
+    serverChannel
+  }
+
+  /**
+   * Listen for new connections and assign accepted connections to processors using round-robin.
+   */
+  private def acceptNewConnections(): Unit = {
+    val ready = nioSelector.select(500)
+    if (ready > 0) {
+      val keys = nioSelector.selectedKeys()
+      val iter = keys.iterator()
+      while (iter.hasNext && isRunning) {
+        try {
+          val key = iter.next
+          iter.remove()
+
+          if (key.isAcceptable) {
+            accept(key).foreach { socketChannel =>
+              // Assign the channel to the next processor (using round-robin) to which the
+              // channel can be added without blocking. If newConnections queue is full on
+              // all processors, block until the last one is able to accept a connection.
+              var retriesLeft = synchronized(processors.length)
+              var processor: Processor = null
+              do {
+                retriesLeft -= 1
+                processor = synchronized {
+                  // adjust the index (if necessary) and retrieve the processor atomically for
+                  // correct behaviour in case the number of processors is reduced dynamically
+                  currentProcessorIndex = currentProcessorIndex % processors.length
+                  processors(currentProcessorIndex)
+                }
+                currentProcessorIndex += 1
+              } while (!assignNewConnection(socketChannel, processor, retriesLeft == 0))
+            }
+          } else
+            throw new IllegalStateException("Unrecognized key state for acceptor thread.")
+        } catch {
+          case e: Throwable => error("Error while accepting connection", e)
+        }
+      }
+    }
+  }
+
+  /**
+   * Accept a new connection
+   */
+  private def accept(key: SelectionKey): Option[SocketChannel] = {
+    val serverSocketChannel = key.channel().asInstanceOf[ServerSocketChannel]
+    val socketChannel = serverSocketChannel.accept()
+    try {
+      connectionQuotas.inc(endPoint.listenerName, socketChannel.socket.getInetAddress, blockedPercentMeter)
+      socketChannel.configureBlocking(false)
+      socketChannel.socket().setTcpNoDelay(true)
+      socketChannel.socket().setKeepAlive(true)
+      if (sendBufferSize != Selectable.USE_DEFAULT_BUFFER_SIZE)
+        socketChannel.socket().setSendBufferSize(sendBufferSize)
+      Some(socketChannel)
+    } catch {
+      case e: TooManyConnectionsException =>
+        info(s"Rejected connection from ${e.ip}, address already has the configured maximum of ${e.count} connections.")
+        close(endPoint.listenerName, socketChannel)
+        None
+      case e: ConnectionThrottledException =>
+        val ip = socketChannel.socket.getInetAddress
+        debug(s"Delaying closing of connection from $ip for ${e.throttleTimeMs} ms")
+        val endThrottleTimeMs = e.startThrottleTimeMs + e.throttleTimeMs
+        throttledSockets += DelayedCloseSocket(socketChannel, endThrottleTimeMs)
+        None
+    }
+  }
+
+  /**
+   * Close sockets for any connections that have been throttled.
+   */
+  private def closeThrottledConnections(): Unit = {
+    val timeMs = time.milliseconds
+    while (throttledSockets.headOption.exists(_.endThrottleTimeMs < timeMs)) {
+      val closingSocket = throttledSockets.dequeue()
+      debug(s"Closing socket from ip ${closingSocket.socket.getRemoteAddress}")
+      closeSocket(closingSocket.socket)
+    }
+  }
+
+  private def assignNewConnection(socketChannel: SocketChannel, processor: Processor, mayBlock: Boolean): Boolean = {
+    if (processor.accept(socketChannel, mayBlock, blockedPercentMeter)) {
+      debug(s"Accepted connection from ${socketChannel.socket.getRemoteSocketAddress} on" +
+        s" ${socketChannel.socket.getLocalSocketAddress} and assigned it to processor ${processor.id}," +
+        s" sendBufferSize [actual|requested]: [${socketChannel.socket.getSendBufferSize}|$sendBufferSize]" +
+        s" recvBufferSize [actual|requested]: [${socketChannel.socket.getReceiveBufferSize}|$recvBufferSize]")
+      true
+    } else
+      false
+  }
+
+  /**
+   * Wakeup the thread for selection.
+   */
+  @Override
+  def wakeup(): Unit = nioSelector.wakeup()
+
+}
