+++ core/src/main/scala/kafka/network/SocketServer.scala
@@ -467,1251 +467,3 @@ object SocketServer {
 
   val ListenerReconfigurableConfigs = Set(KafkaConfig.MaxConnectionsProp, KafkaConfig.MaxConnectionCreationRateProp)
 }
-
-/**
- * A base class with some helper variables and methods
- */
-private[kafka] abstract class AbstractServerThread(connectionQuotas: ConnectionQuotas) extends Runnable with Logging {
-
-  private val startupLatch = new CountDownLatch(1)
-
-  // `shutdown()` is invoked before `startupComplete` and `shutdownComplete` if an exception is thrown in the constructor
-  // (e.g. if the address is already in use). We want `shutdown` to proceed in such cases, so we first assign an open
-  // latch and then replace it in `startupComplete()`.
-  @volatile private var shutdownLatch = new CountDownLatch(0)
-
-  private val alive = new AtomicBoolean(true)
-
-  def wakeup(): Unit
-
-  /**
-   * Initiates a graceful shutdown by signaling to stop
-   */
-  def initiateShutdown(): Unit = {
-    if (alive.getAndSet(false))
-      wakeup()
-  }
-
-  /**
-   * Wait for the thread to completely shutdown
-   */
-  def awaitShutdown(): Unit = shutdownLatch.await
-
-  /**
-   * Returns true if the thread is completely started
-   */
-  def isStarted(): Boolean = startupLatch.getCount == 0
-
-  /**
-   * Wait for the thread to completely start up
-   */
-  def awaitStartup(): Unit = startupLatch.await
-
-  /**
-   * Record that the thread startup is complete
-   */
-  protected def startupComplete(): Unit = {
-    // Replace the open latch with a closed one
-    shutdownLatch = new CountDownLatch(1)
-    startupLatch.countDown()
-  }
-
-  /**
-   * Record that the thread shutdown is complete
-   */
-  protected def shutdownComplete(): Unit = shutdownLatch.countDown()
-
-  /**
-   * Is the server still running?
-   */
-  protected def isRunning: Boolean = alive.get
-
-  /**
-   * Close `channel` and decrement the connection count.
-   */
-  def close(listenerName: ListenerName, channel: SocketChannel): Unit = {
-    if (channel != null) {
-      debug(s"Closing connection from ${channel.socket.getRemoteSocketAddress}")
-      connectionQuotas.dec(listenerName, channel.socket.getInetAddress)
-      closeSocket(channel)
-    }
-  }
-
-  protected def closeSocket(channel: SocketChannel): Unit = {
-    CoreUtils.swallow(channel.socket().close(), this, Level.ERROR)
-    CoreUtils.swallow(channel.close(), this, Level.ERROR)
-  }
-}
-
-/**
- * Thread that accepts and configures new connections. There is one of these per endpoint.
- */
-private[kafka] class Acceptor(val endPoint: EndPoint,
-                              val sendBufferSize: Int,
-                              val recvBufferSize: Int,
-                              nodeId: Int,
-                              connectionQuotas: ConnectionQuotas,
-                              metricPrefix: String,
-                              time: Time,
-                              logPrefix: String = "") extends AbstractServerThread(connectionQuotas) with KafkaMetricsGroup {
-
-  this.logIdent = logPrefix
-  private val nioSelector = NSelector.open()
-  val serverChannel = openServerSocket(endPoint.host, endPoint.port)
-  private val processors = new ArrayBuffer[Processor]()
-  private val processorsStarted = new AtomicBoolean
-  private val blockedPercentMeter = newMeter(s"${metricPrefix}AcceptorBlockedPercent",
-    "blocked time", TimeUnit.NANOSECONDS, Map(ListenerMetricTag -> endPoint.listenerName.value))
-  private var currentProcessorIndex = 0
-  private[network] val throttledSockets = new mutable.PriorityQueue[DelayedCloseSocket]()
-
-  private[network] case class DelayedCloseSocket(socket: SocketChannel, endThrottleTimeMs: Long) extends Ordered[DelayedCloseSocket] {
-    override def compare(that: DelayedCloseSocket): Int = endThrottleTimeMs compare that.endThrottleTimeMs
-  }
-
-  private[network] def addProcessors(newProcessors: Buffer[Processor], processorThreadPrefix: String): Unit = synchronized {
-    processors ++= newProcessors
-    if (processorsStarted.get)
-      startProcessors(newProcessors, processorThreadPrefix)
-  }
-
-  private[network] def startProcessors(processorThreadPrefix: String): Unit = synchronized {
-    if (!processorsStarted.getAndSet(true)) {
-      startProcessors(processors, processorThreadPrefix)
-    }
-  }
-
-  private def startProcessors(processors: Seq[Processor], processorThreadPrefix: String): Unit = synchronized {
-    processors.foreach { processor =>
-      KafkaThread.nonDaemon(
-        s"${processorThreadPrefix}-kafka-network-thread-$nodeId-${endPoint.listenerName}-${endPoint.securityProtocol}-${processor.id}",
-        processor
-      ).start()
-    }
-  }
-
-  private[network] def removeProcessors(removeCount: Int, requestChannel: RequestChannel): Unit = synchronized {
-    // Shutdown `removeCount` processors. Remove them from the processor list first so that no more
-    // connections are assigned. Shutdown the removed processors, closing the selector and its connections.
-    // The processors are then removed from `requestChannel` and any pending responses to these processors are dropped.
-    val toRemove = processors.takeRight(removeCount)
-    processors.remove(processors.size - removeCount, removeCount)
-    toRemove.foreach(_.initiateShutdown())
-    toRemove.foreach(_.awaitShutdown())
-    toRemove.foreach(processor => requestChannel.removeProcessor(processor.id))
-  }
-
-  override def initiateShutdown(): Unit = {
-    super.initiateShutdown()
-    synchronized {
-      processors.foreach(_.initiateShutdown())
-    }
-  }
-
-  override def awaitShutdown(): Unit = {
-    super.awaitShutdown()
-    synchronized {
-      processors.foreach(_.awaitShutdown())
-    }
-  }
-
-  /**
-   * Accept loop that checks for new connection attempts
-   */
-  def run(): Unit = {
-    serverChannel.register(nioSelector, SelectionKey.OP_ACCEPT)
-    startupComplete()
-    try {
-      while (isRunning) {
-        try {
-          acceptNewConnections()
-          closeThrottledConnections()
-        }
-        catch {
-          // We catch all the throwables to prevent the acceptor thread from exiting on exceptions due
-          // to a select operation on a specific channel or a bad request. We don't want
-          // the broker to stop responding to requests from other clients in these scenarios.
-          case e: ControlThrowable => throw e
-          case e: Throwable => error("Error occurred", e)
-        }
-      }
-    } finally {
-      debug("Closing server socket, selector, and any throttled sockets.")
-      CoreUtils.swallow(serverChannel.close(), this, Level.ERROR)
-      CoreUtils.swallow(nioSelector.close(), this, Level.ERROR)
-      throttledSockets.foreach(throttledSocket => closeSocket(throttledSocket.socket))
-      throttledSockets.clear()
-      shutdownComplete()
-    }
-  }
-
-  /**
-   * Create a server socket to listen for connections on.
-   */
-  private def openServerSocket(host: String, port: Int): ServerSocketChannel = {
-    val socketAddress =
-      if (Utils.isBlank(host))
-        new InetSocketAddress(port)
-      else
-        new InetSocketAddress(host, port)
-    val serverChannel = ServerSocketChannel.open()
-    serverChannel.configureBlocking(false)
-    if (recvBufferSize != Selectable.USE_DEFAULT_BUFFER_SIZE)
-      serverChannel.socket().setReceiveBufferSize(recvBufferSize)
-
-    try {
-      serverChannel.socket.bind(socketAddress)
-      info(s"Awaiting socket connections on ${socketAddress.getHostString}:${serverChannel.socket.getLocalPort}.")
-    } catch {
-      case e: SocketException =>
-        throw new KafkaException(s"Socket server failed to bind to ${socketAddress.getHostString}:$port: ${e.getMessage}.", e)
-    }
-    serverChannel
-  }
-
-  /**
-   * Listen for new connections and assign accepted connections to processors using round-robin.
-   */
-  private def acceptNewConnections(): Unit = {
-    val ready = nioSelector.select(500)
-    if (ready > 0) {
-      val keys = nioSelector.selectedKeys()
-      val iter = keys.iterator()
-      while (iter.hasNext && isRunning) {
-        try {
-          val key = iter.next
-          iter.remove()
-
-          if (key.isAcceptable) {
-            accept(key).foreach { socketChannel =>
-              // Assign the channel to the next processor (using round-robin) to which the
-              // channel can be added without blocking. If newConnections queue is full on
-              // all processors, block until the last one is able to accept a connection.
-              var retriesLeft = synchronized(processors.length)
-              var processor: Processor = null
-              do {
-                retriesLeft -= 1
-                processor = synchronized {
-                  // adjust the index (if necessary) and retrieve the processor atomically for
-                  // correct behaviour in case the number of processors is reduced dynamically
-                  currentProcessorIndex = currentProcessorIndex % processors.length
-                  processors(currentProcessorIndex)
-                }
-                currentProcessorIndex += 1
-              } while (!assignNewConnection(socketChannel, processor, retriesLeft == 0))
-            }
-          } else
-            throw new IllegalStateException("Unrecognized key state for acceptor thread.")
-        } catch {
-          case e: Throwable => error("Error while accepting connection", e)
-        }
-      }
-    }
-  }
-
-  /**
-   * Accept a new connection
-   */
-  private def accept(key: SelectionKey): Option[SocketChannel] = {
-    val serverSocketChannel = key.channel().asInstanceOf[ServerSocketChannel]
-    val socketChannel = serverSocketChannel.accept()
-    try {
-      connectionQuotas.inc(endPoint.listenerName, socketChannel.socket.getInetAddress, blockedPercentMeter)
-      socketChannel.configureBlocking(false)
-      socketChannel.socket().setTcpNoDelay(true)
-      socketChannel.socket().setKeepAlive(true)
-      if (sendBufferSize != Selectable.USE_DEFAULT_BUFFER_SIZE)
-        socketChannel.socket().setSendBufferSize(sendBufferSize)
-      Some(socketChannel)
-    } catch {
-      case e: TooManyConnectionsException =>
-        info(s"Rejected connection from ${e.ip}, address already has the configured maximum of ${e.count} connections.")
-        close(endPoint.listenerName, socketChannel)
-        None
-      case e: ConnectionThrottledException =>
-        val ip = socketChannel.socket.getInetAddress
-        debug(s"Delaying closing of connection from $ip for ${e.throttleTimeMs} ms")
-        val endThrottleTimeMs = e.startThrottleTimeMs + e.throttleTimeMs
-        throttledSockets += DelayedCloseSocket(socketChannel, endThrottleTimeMs)
-        None
-    }
-  }
-
-  /**
-   * Close sockets for any connections that have been throttled.
-   */
-  private def closeThrottledConnections(): Unit = {
-    val timeMs = time.milliseconds
-    while (throttledSockets.headOption.exists(_.endThrottleTimeMs < timeMs)) {
-      val closingSocket = throttledSockets.dequeue()
-      debug(s"Closing socket from ip ${closingSocket.socket.getRemoteAddress}")
-      closeSocket(closingSocket.socket)
-    }
-  }
-
-  private def assignNewConnection(socketChannel: SocketChannel, processor: Processor, mayBlock: Boolean): Boolean = {
-    if (processor.accept(socketChannel, mayBlock, blockedPercentMeter)) {
-      debug(s"Accepted connection from ${socketChannel.socket.getRemoteSocketAddress} on" +
-        s" ${socketChannel.socket.getLocalSocketAddress} and assigned it to processor ${processor.id}," +
-        s" sendBufferSize [actual|requested]: [${socketChannel.socket.getSendBufferSize}|$sendBufferSize]" +
-        s" recvBufferSize [actual|requested]: [${socketChannel.socket.getReceiveBufferSize}|$recvBufferSize]")
-      true
-    } else
-      false
-  }
-
-  /**
-   * Wakeup the thread for selection.
-   */
-  @Override
-  def wakeup(): Unit = nioSelector.wakeup()
-
-}
-
-private[kafka] object Processor {
-  val IdlePercentMetricName = "IdlePercent"
-  val NetworkProcessorMetricTag = "networkProcessor"
-  val ListenerMetricTag = "listener"
-  val ConnectionQueueSize = 20
-}
-
-/**
- * Thread that processes all requests from a single connection. There are N of these running in parallel
- * each of which has its own selector
- *
- * @param isPrivilegedListener The privileged listener flag is used as one factor to determine whether
- *                             a certain request is forwarded or not. When the control plane is defined,
- *                             the control plane processor would be fellow broker's choice for sending
- *                             forwarding requests; if the control plane is not defined, the processor
- *                             relying on the inter broker listener would be acting as the privileged listener.
- */
-private[kafka] class Processor(val id: Int,
-                               time: Time,
-                               maxRequestSize: Int,
-                               requestChannel: RequestChannel,
-                               connectionQuotas: ConnectionQuotas,
-                               connectionsMaxIdleMs: Long,
-                               failedAuthenticationDelayMs: Int,
-                               listenerName: ListenerName,
-                               securityProtocol: SecurityProtocol,
-                               config: KafkaConfig,
-                               metrics: Metrics,
-                               credentialProvider: CredentialProvider,
-                               memoryPool: MemoryPool,
-                               logContext: LogContext,
-                               connectionQueueSize: Int,
-                               isPrivilegedListener: Boolean,
-                               apiVersionManager: ApiVersionManager) extends AbstractServerThread(connectionQuotas) with KafkaMetricsGroup {
-
-  private object ConnectionId {
-    def fromString(s: String): Option[ConnectionId] = s.split("-") match {
-      case Array(local, remote, index) => BrokerEndPoint.parseHostPort(local).flatMap { case (localHost, localPort) =>
-        BrokerEndPoint.parseHostPort(remote).map { case (remoteHost, remotePort) =>
-          ConnectionId(localHost, localPort, remoteHost, remotePort, Integer.parseInt(index))
-        }
-      }
-      case _ => None
-    }
-  }
-
-  private[network] case class ConnectionId(localHost: String, localPort: Int, remoteHost: String, remotePort: Int, index: Int) {
-    override def toString: String = s"$localHost:$localPort-$remoteHost:$remotePort-$index"
-  }
-
-  private val newConnections = new ArrayBlockingQueue[SocketChannel](connectionQueueSize)
-  private val inflightResponses = mutable.Map[String, RequestChannel.Response]()
-  private val responseQueue = new LinkedBlockingDeque[RequestChannel.Response]()
-
-  private[kafka] val metricTags = mutable.LinkedHashMap(
-    ListenerMetricTag -> listenerName.value,
-    NetworkProcessorMetricTag -> id.toString
-  ).asJava
-
-  newGauge(IdlePercentMetricName, () => {
-    Option(metrics.metric(metrics.metricName("io-wait-ratio", MetricsGroup, metricTags))).fold(0.0)(m =>
-      Math.min(m.metricValue.asInstanceOf[Double], 1.0))
-  },
-    // for compatibility, only add a networkProcessor tag to the Yammer Metrics alias (the equivalent Selector metric
-    // also includes the listener name)
-    Map(NetworkProcessorMetricTag -> id.toString)
-  )
-
-  val expiredConnectionsKilledCount = new CumulativeSum()
-  private val expiredConnectionsKilledCountMetricName = metrics.metricName("expired-connections-killed-count", MetricsGroup, metricTags)
-  metrics.addMetric(expiredConnectionsKilledCountMetricName, expiredConnectionsKilledCount)
-
-  private val selector = createSelector(
-    ChannelBuilders.serverChannelBuilder(
-      listenerName,
-      listenerName == config.interBrokerListenerName,
-      securityProtocol,
-      config,
-      credentialProvider.credentialCache,
-      credentialProvider.tokenCache,
-      time,
-      logContext,
-      () => apiVersionManager.apiVersionResponse(throttleTimeMs = 0)
-    )
-  )
-
-  // Visible to override for testing
-  protected[network] def createSelector(channelBuilder: ChannelBuilder): KSelector = {
-    channelBuilder match {
-      case reconfigurable: Reconfigurable => config.addReconfigurable(reconfigurable)
-      case _ =>
-    }
-    new KSelector(
-      maxRequestSize,
-      connectionsMaxIdleMs,
-      failedAuthenticationDelayMs,
-      metrics,
-      time,
-      "socket-server",
-      metricTags,
-      false,
-      true,
-      channelBuilder,
-      memoryPool,
-      logContext)
-  }
-
-  // Connection ids have the format `localAddr:localPort-remoteAddr:remotePort-index`. The index is a
-  // non-negative incrementing value that ensures that even if remotePort is reused after a connection is
-  // closed, connection ids are not reused while requests from the closed connection are being processed.
-  private var nextConnectionIndex = 0
-
-  override def run(): Unit = {
-    startupComplete()
-    try {
-      while (isRunning) {
-        try {
-          // setup any new connections that have been queued up
-          configureNewConnections()
-          // register any new responses for writing
-          processNewResponses()
-          poll()
-          processCompletedReceives()
-          processCompletedSends()
-          processDisconnected()
-          closeExcessConnections()
-        } catch {
-          // We catch all the throwables here to prevent the processor thread from exiting. We do this because
-          // letting a processor exit might cause a bigger impact on the broker. This behavior might need to be
-          // reviewed if we see an exception that needs the entire broker to stop. Usually the exceptions thrown would
-          // be either associated with a specific socket channel or a bad request. These exceptions are caught and
-          // processed by the individual methods above which close the failing channel and continue processing other
-          // channels. So this catch block should only ever see ControlThrowables.
-          case e: Throwable => processException("Processor got uncaught exception.", e)
-        }
-      }
-    } finally {
-      debug(s"Closing selector - processor $id")
-      CoreUtils.swallow(closeAll(), this, Level.ERROR)
-      shutdownComplete()
-    }
-  }
-
-  private[network] def processException(errorMessage: String, throwable: Throwable): Unit = {
-    throwable match {
-      case e: ControlThrowable => throw e
-      case e => error(errorMessage, e)
-    }
-  }
-
-  private def processChannelException(channelId: String, errorMessage: String, throwable: Throwable): Unit = {
-    if (openOrClosingChannel(channelId).isDefined) {
-      error(s"Closing socket for $channelId because of error", throwable)
-      close(channelId)
-    }
-    processException(errorMessage, throwable)
-  }
-
-  private def processNewResponses(): Unit = {
-    var currentResponse: RequestChannel.Response = null
-    while ({currentResponse = dequeueResponse(); currentResponse != null}) {
-      val channelId = currentResponse.request.context.connectionId
-      try {
-        currentResponse match {
-          case response: NoOpResponse =>
-            // There is no response to send to the client, we need to read more pipelined requests
-            // that are sitting in the server's socket buffer
-            updateRequestMetrics(response)
-            trace(s"Socket server received empty response to send, registering for read: $response")
-            // Try unmuting the channel. If there was no quota violation and the channel has not been throttled,
-            // it will be unmuted immediately. If the channel has been throttled, it will be unmuted only if the
-            // throttling delay has already passed by now.
-            handleChannelMuteEvent(channelId, ChannelMuteEvent.RESPONSE_SENT)
-            tryUnmuteChannel(channelId)
-
-          case response: SendResponse =>
-            sendResponse(response, response.responseSend)
-          case response: CloseConnectionResponse =>
-            updateRequestMetrics(response)
-            trace("Closing socket connection actively according to the response code.")
-            close(channelId)
-          case _: StartThrottlingResponse =>
-            handleChannelMuteEvent(channelId, ChannelMuteEvent.THROTTLE_STARTED)
-          case _: EndThrottlingResponse =>
-            // Try unmuting the channel. The channel will be unmuted only if the response has already been sent out to
-            // the client.
-            handleChannelMuteEvent(channelId, ChannelMuteEvent.THROTTLE_ENDED)
-            tryUnmuteChannel(channelId)
-          case _ =>
-            throw new IllegalArgumentException(s"Unknown response type: ${currentResponse.getClass}")
-        }
-      } catch {
-        case e: Throwable =>
-          processChannelException(channelId, s"Exception while processing response for $channelId", e)
-      }
-    }
-  }
-
-  // `protected` for test usage
-  protected[network] def sendResponse(response: RequestChannel.Response, responseSend: Send): Unit = {
-    val connectionId = response.request.context.connectionId
-    trace(s"Socket server received response to send to $connectionId, registering for write and sending data: $response")
-    // `channel` can be None if the connection was closed remotely or if selector closed it for being idle for too long
-    if (channel(connectionId).isEmpty) {
-      warn(s"Attempting to send response via channel for which there is no open connection, connection id $connectionId")
-      response.request.updateRequestMetrics(0L, response)
-    }
-    // Invoke send for closingChannel as well so that the send is failed and the channel closed properly and
-    // removed from the Selector after discarding any pending staged receives.
-    // `openOrClosingChannel` can be None if the selector closed the connection because it was idle for too long
-    if (openOrClosingChannel(connectionId).isDefined) {
-      selector.send(new NetworkSend(connectionId, responseSend))
-      inflightResponses += (connectionId -> response)
-    }
-  }
-
-  private def poll(): Unit = {
-    val pollTimeout = if (newConnections.isEmpty) 300 else 0
-    try selector.poll(pollTimeout)
-    catch {
-      case e @ (_: IllegalStateException | _: IOException) =>
-        // The exception is not re-thrown and any completed sends/receives/connections/disconnections
-        // from this poll will be processed.
-        error(s"Processor $id poll failed", e)
-    }
-  }
-
-  protected def parseRequestHeader(buffer: ByteBuffer): RequestHeader = {
-    val header = RequestHeader.parse(buffer)
-    if (apiVersionManager.isApiEnabled(header.apiKey)) {
-      header
-    } else {
-      throw new InvalidRequestException(s"Received request api key ${header.apiKey} which is not enabled")
-    }
-  }
-
-  private def processCompletedReceives(): Unit = {
-    selector.completedReceives.forEach { receive =>
-      try {
-        openOrClosingChannel(receive.source) match {
-          case Some(channel) =>
-            val header = parseRequestHeader(receive.payload)
-            if (header.apiKey == ApiKeys.SASL_HANDSHAKE && channel.maybeBeginServerReauthentication(receive,
-              () => time.nanoseconds()))
-              trace(s"Begin re-authentication: $channel")
-            else {
-              val nowNanos = time.nanoseconds()
-              if (channel.serverAuthenticationSessionExpired(nowNanos)) {
-                // be sure to decrease connection count and drop any in-flight responses
-                debug(s"Disconnecting expired channel: $channel : $header")
-                close(channel.id)
-                expiredConnectionsKilledCount.record(null, 1, 0)
-              } else {
-                val connectionId = receive.source
-                val context = new RequestContext(header, connectionId, channel.socketAddress,
-                  channel.principal, listenerName, securityProtocol,
-                  channel.channelMetadataRegistry.clientInformation, isPrivilegedListener, channel.principalSerde)
-
-                val req = new RequestChannel.Request(processor = id, context = context,
-                  startTimeNanos = nowNanos, memoryPool, receive.payload, requestChannel.metrics, None)
-
-                // KIP-511: ApiVersionsRequest is intercepted here to catch the client software name
-                // and version. It is done here to avoid wiring things up to the api layer.
-                if (header.apiKey == ApiKeys.API_VERSIONS) {
-                  val apiVersionsRequest = req.body[ApiVersionsRequest]
-                  if (apiVersionsRequest.isValid) {
-                    channel.channelMetadataRegistry.registerClientInformation(new ClientInformation(
-                      apiVersionsRequest.data.clientSoftwareName,
-                      apiVersionsRequest.data.clientSoftwareVersion))
-                  }
-                }
-                requestChannel.sendRequest(req)
-                selector.mute(connectionId)
-                handleChannelMuteEvent(connectionId, ChannelMuteEvent.REQUEST_RECEIVED)
-              }
-            }
-          case None =>
-            // This should never happen since completed receives are processed immediately after `poll()`
-            throw new IllegalStateException(s"Channel ${receive.source} removed from selector before processing completed receive")
-        }
-      } catch {
-        // note that even though we got an exception, we can assume that receive.source is valid.
-        // Issues with constructing a valid receive object were handled earlier
-        case e: Throwable =>
-          processChannelException(receive.source, s"Exception while processing request from ${receive.source}", e)
-      }
-    }
-    selector.clearCompletedReceives()
-  }
-
-  private def processCompletedSends(): Unit = {
-    selector.completedSends.forEach { send =>
-      try {
-        val response = inflightResponses.remove(send.destinationId).getOrElse {
-          throw new IllegalStateException(s"Send for ${send.destinationId} completed, but not in `inflightResponses`")
-        }
-        updateRequestMetrics(response)
-
-        // Invoke send completion callback
-        response.onComplete.foreach(onComplete => onComplete(send))
-
-        // Try unmuting the channel. If there was no quota violation and the channel has not been throttled,
-        // it will be unmuted immediately. If the channel has been throttled, it will unmuted only if the throttling
-        // delay has already passed by now.
-        handleChannelMuteEvent(send.destinationId, ChannelMuteEvent.RESPONSE_SENT)
-        tryUnmuteChannel(send.destinationId)
-      } catch {
-        case e: Throwable => processChannelException(send.destinationId,
-          s"Exception while processing completed send to ${send.destinationId}", e)
-      }
-    }
-    selector.clearCompletedSends()
-  }
-
-  private def updateRequestMetrics(response: RequestChannel.Response): Unit = {
-    val request = response.request
-    val networkThreadTimeNanos = openOrClosingChannel(request.context.connectionId).fold(0L)(_.getAndResetNetworkThreadTimeNanos())
-    request.updateRequestMetrics(networkThreadTimeNanos, response)
-  }
-
-  private def processDisconnected(): Unit = {
-    selector.disconnected.keySet.forEach { connectionId =>
-      try {
-        val remoteHost = ConnectionId.fromString(connectionId).getOrElse {
-          throw new IllegalStateException(s"connectionId has unexpected format: $connectionId")
-        }.remoteHost
-        inflightResponses.remove(connectionId).foreach(updateRequestMetrics)
-        // the channel has been closed by the selector but the quotas still need to be updated
-        connectionQuotas.dec(listenerName, InetAddress.getByName(remoteHost))
-      } catch {
-        case e: Throwable => processException(s"Exception while processing disconnection of $connectionId", e)
-      }
-    }
-  }
-
-  private def closeExcessConnections(): Unit = {
-    if (connectionQuotas.maxConnectionsExceeded(listenerName)) {
-      val channel = selector.lowestPriorityChannel()
-      if (channel != null)
-        close(channel.id)
-    }
-  }
-
-  /**
-   * Close the connection identified by `connectionId` and decrement the connection count.
-   * The channel will be immediately removed from the selector's `channels` or `closingChannels`
-   * and no further disconnect notifications will be sent for this channel by the selector.
-   * If responses are pending for the channel, they are dropped and metrics is updated.
-   * If the channel has already been removed from selector, no action is taken.
-   */
-  private def close(connectionId: String): Unit = {
-    openOrClosingChannel(connectionId).foreach { channel =>
-      debug(s"Closing selector connection $connectionId")
-      val address = channel.socketAddress
-      if (address != null)
-        connectionQuotas.dec(listenerName, address)
-      selector.close(connectionId)
-
-      inflightResponses.remove(connectionId).foreach(response => updateRequestMetrics(response))
-    }
-  }
-
-  /**
-   * Queue up a new connection for reading
-   */
-  def accept(socketChannel: SocketChannel,
-             mayBlock: Boolean,
-             acceptorIdlePercentMeter: com.yammer.metrics.core.Meter): Boolean = {
-    val accepted = {
-      if (newConnections.offer(socketChannel))
-        true
-      else if (mayBlock) {
-        val startNs = time.nanoseconds
-        newConnections.put(socketChannel)
-        acceptorIdlePercentMeter.mark(time.nanoseconds() - startNs)
-        true
-      } else
-        false
-    }
-    if (accepted)
-      wakeup()
-    accepted
-  }
-
-  /**
-   * Register any new connections that have been queued up. The number of connections processed
-   * in each iteration is limited to ensure that traffic and connection close notifications of
-   * existing channels are handled promptly.
-   */
-  private def configureNewConnections(): Unit = {
-    var connectionsProcessed = 0
-    while (connectionsProcessed < connectionQueueSize && !newConnections.isEmpty) {
-      val channel = newConnections.poll()
-      try {
-        debug(s"Processor $id listening to new connection from ${channel.socket.getRemoteSocketAddress}")
-        selector.register(connectionId(channel.socket), channel)
-        connectionsProcessed += 1
-      } catch {
-        // We explicitly catch all exceptions and close the socket to avoid a socket leak.
-        case e: Throwable =>
-          val remoteAddress = channel.socket.getRemoteSocketAddress
-          // need to close the channel here to avoid a socket leak.
-          close(listenerName, channel)
-          processException(s"Processor $id closed connection from $remoteAddress", e)
-      }
-    }
-  }
-
-  /**
-   * Close the selector and all open connections
-   */
-  private def closeAll(): Unit = {
-    while (!newConnections.isEmpty) {
-      newConnections.poll().close()
-    }
-    selector.channels.forEach { channel =>
-      close(channel.id)
-    }
-    selector.close()
-    removeMetric(IdlePercentMetricName, Map(NetworkProcessorMetricTag -> id.toString))
-  }
-
-  // 'protected` to allow override for testing
-  protected[network] def connectionId(socket: Socket): String = {
-    val localHost = socket.getLocalAddress.getHostAddress
-    val localPort = socket.getLocalPort
-    val remoteHost = socket.getInetAddress.getHostAddress
-    val remotePort = socket.getPort
-    val connId = ConnectionId(localHost, localPort, remoteHost, remotePort, nextConnectionIndex).toString
-    nextConnectionIndex = if (nextConnectionIndex == Int.MaxValue) 0 else nextConnectionIndex + 1
-    connId
-  }
-
-  private[network] def enqueueResponse(response: RequestChannel.Response): Unit = {
-    responseQueue.put(response)
-    wakeup()
-  }
-
-  private def dequeueResponse(): RequestChannel.Response = {
-    val response = responseQueue.poll()
-    if (response != null)
-      response.request.responseDequeueTimeNanos = Time.SYSTEM.nanoseconds
-    response
-  }
-
-  private[network] def responseQueueSize = responseQueue.size
-
-  // Only for testing
-  private[network] def inflightResponseCount: Int = inflightResponses.size
-
-  // Visible for testing
-  // Only methods that are safe to call on a disconnected channel should be invoked on 'openOrClosingChannel'.
-  private[network] def openOrClosingChannel(connectionId: String): Option[KafkaChannel] =
-    Option(selector.channel(connectionId)).orElse(Option(selector.closingChannel(connectionId)))
-
-  // Indicate the specified channel that the specified channel mute-related event has happened so that it can change its
-  // mute state.
-  private def handleChannelMuteEvent(connectionId: String, event: ChannelMuteEvent): Unit = {
-    openOrClosingChannel(connectionId).foreach(c => c.handleChannelMuteEvent(event))
-  }
-
-  private def tryUnmuteChannel(connectionId: String) = {
-    openOrClosingChannel(connectionId).foreach(c => selector.unmute(c.id))
-  }
-
-  /* For test usage */
-  private[network] def channel(connectionId: String): Option[KafkaChannel] =
-    Option(selector.channel(connectionId))
-
-  /**
-   * Wakeup the thread for selection.
-   */
-  override def wakeup() = selector.wakeup()
-
-  override def initiateShutdown(): Unit = {
-    super.initiateShutdown()
-    removeMetric("IdlePercent", Map("networkProcessor" -> id.toString))
-    metrics.removeMetric(expiredConnectionsKilledCountMetricName)
-  }
-}
-
-/**
- * Interface for connection quota configuration. Connection quotas can be configured at the
- * broker, listener or IP level.
- */
-sealed trait ConnectionQuotaEntity {
-  def sensorName: String
-  def metricName: String
-  def sensorExpiration: Long
-  def metricTags: Map[String, String]
-}
-
-object ConnectionQuotas {
-  private val InactiveSensorExpirationTimeSeconds = TimeUnit.HOURS.toSeconds(1)
-  private val ConnectionRateSensorName = "Connection-Accept-Rate"
-  private val ConnectionRateMetricName = "connection-accept-rate"
-  private val IpMetricTag = "ip"
-  private val ListenerThrottlePrefix = ""
-  private val IpThrottlePrefix = "ip-"
-
-  private case class ListenerQuotaEntity(listenerName: String) extends ConnectionQuotaEntity {
-    override def sensorName: String = s"$ConnectionRateSensorName-$listenerName"
-    override def sensorExpiration: Long = Long.MaxValue
-    override def metricName: String = ConnectionRateMetricName
-    override def metricTags: Map[String, String] = Map(ListenerMetricTag -> listenerName)
-  }
-
-  private case object BrokerQuotaEntity extends ConnectionQuotaEntity {
-    override def sensorName: String = ConnectionRateSensorName
-    override def sensorExpiration: Long = Long.MaxValue
-    override def metricName: String = s"broker-$ConnectionRateMetricName"
-    override def metricTags: Map[String, String] = Map.empty
-  }
-
-  private case class IpQuotaEntity(ip: InetAddress) extends ConnectionQuotaEntity {
-    override def sensorName: String = s"$ConnectionRateSensorName-${ip.getHostAddress}"
-    override def sensorExpiration: Long = InactiveSensorExpirationTimeSeconds
-    override def metricName: String = ConnectionRateMetricName
-    override def metricTags: Map[String, String] = Map(IpMetricTag -> ip.getHostAddress)
-  }
-}
-
-class ConnectionQuotas(config: KafkaConfig, time: Time, metrics: Metrics) extends Logging with AutoCloseable {
-
-  @volatile private var defaultMaxConnectionsPerIp: Int = config.maxConnectionsPerIp
-  @volatile private var maxConnectionsPerIpOverrides = config.maxConnectionsPerIpOverrides.map { case (host, count) => (InetAddress.getByName(host), count) }
-  @volatile private var brokerMaxConnections = config.maxConnections
-  private val interBrokerListenerName = config.interBrokerListenerName
-  private val counts = mutable.Map[InetAddress, Int]()
-
-  // Listener counts and configs are synchronized on `counts`
-  private val listenerCounts = mutable.Map[ListenerName, Int]()
-  private[network] val maxConnectionsPerListener = mutable.Map[ListenerName, ListenerConnectionQuota]()
-  @volatile private var totalCount = 0
-  // updates to defaultConnectionRatePerIp or connectionRatePerIp must be synchronized on `counts`
-  @volatile private var defaultConnectionRatePerIp = QuotaConfigs.IP_CONNECTION_RATE_DEFAULT.intValue()
-  private val connectionRatePerIp = new ConcurrentHashMap[InetAddress, Int]()
-  // sensor that tracks broker-wide connection creation rate and limit (quota)
-  private val brokerConnectionRateSensor = getOrCreateConnectionRateQuotaSensor(config.maxConnectionCreationRate, BrokerQuotaEntity)
-  private val maxThrottleTimeMs = TimeUnit.SECONDS.toMillis(config.quotaWindowSizeSeconds.toLong)
-
-  def inc(listenerName: ListenerName, address: InetAddress, acceptorBlockedPercentMeter: com.yammer.metrics.core.Meter): Unit = {
-    counts.synchronized {
-      waitForConnectionSlot(listenerName, acceptorBlockedPercentMeter)
-
-      recordIpConnectionMaybeThrottle(listenerName, address)
-      val count = counts.getOrElseUpdate(address, 0)
-      counts.put(address, count + 1)
-      totalCount += 1
-      if (listenerCounts.contains(listenerName)) {
-        listenerCounts.put(listenerName, listenerCounts(listenerName) + 1)
-      }
-      val max = maxConnectionsPerIpOverrides.getOrElse(address, defaultMaxConnectionsPerIp)
-      if (count >= max)
-        throw new TooManyConnectionsException(address, max)
-    }
-  }
-
-  private[network] def updateMaxConnectionsPerIp(maxConnectionsPerIp: Int): Unit = {
-    defaultMaxConnectionsPerIp = maxConnectionsPerIp
-  }
-
-  private[network] def updateMaxConnectionsPerIpOverride(overrideQuotas: Map[String, Int]): Unit = {
-    maxConnectionsPerIpOverrides = overrideQuotas.map { case (host, count) => (InetAddress.getByName(host), count) }
-  }
-
-  private[network] def updateBrokerMaxConnections(maxConnections: Int): Unit = {
-    counts.synchronized {
-      brokerMaxConnections = maxConnections
-      counts.notifyAll()
-    }
-  }
-
-  private[network] def updateBrokerMaxConnectionRate(maxConnectionRate: Int): Unit = {
-    // if there is a connection waiting on the rate throttle delay, we will let it wait the original delay even if
-    // the rate limit increases, because it is just one connection per listener and the code is simpler that way
-    updateConnectionRateQuota(maxConnectionRate, BrokerQuotaEntity)
-  }
-
-  /**
-   * Update the connection rate quota for a given IP and updates quota configs for updated IPs.
-   * If an IP is given, metric config will be updated only for the given IP, otherwise
-   * all metric configs will be checked and updated if required.
-   *
-   * @param ip ip to update or default if None
-   * @param maxConnectionRate new connection rate, or resets entity to default if None
-   */
-  def updateIpConnectionRateQuota(ip: Option[InetAddress], maxConnectionRate: Option[Int]): Unit = synchronized {
-    def isIpConnectionRateMetric(metricName: MetricName) = {
-      metricName.name == ConnectionRateMetricName &&
-      metricName.group == MetricsGroup &&
-      metricName.tags.containsKey(IpMetricTag)
-    }
-
-    def shouldUpdateQuota(metric: KafkaMetric, quotaLimit: Int) = {
-      quotaLimit != metric.config.quota.bound
-    }
-
-    ip match {
-      case Some(address) =>
-        // synchronize on counts to ensure reading an IP connection rate quota and creating a quota config is atomic
-        counts.synchronized {
-          maxConnectionRate match {
-            case Some(rate) =>
-              info(s"Updating max connection rate override for $address to $rate")
-              connectionRatePerIp.put(address, rate)
-            case None =>
-              info(s"Removing max connection rate override for $address")
-              connectionRatePerIp.remove(address)
-          }
-        }
-        updateConnectionRateQuota(connectionRateForIp(address), IpQuotaEntity(address))
-      case None =>
-        // synchronize on counts to ensure reading an IP connection rate quota and creating a quota config is atomic
-        counts.synchronized {
-          defaultConnectionRatePerIp = maxConnectionRate.getOrElse(QuotaConfigs.IP_CONNECTION_RATE_DEFAULT.intValue())
-        }
-        info(s"Updated default max IP connection rate to $defaultConnectionRatePerIp")
-        metrics.metrics.forEach { (metricName, metric) =>
-          if (isIpConnectionRateMetric(metricName)) {
-            val quota = connectionRateForIp(InetAddress.getByName(metricName.tags.get(IpMetricTag)))
-            if (shouldUpdateQuota(metric, quota)) {
-              debug(s"Updating existing connection rate quota config for ${metricName.tags} to $quota")
-              metric.config(rateQuotaMetricConfig(quota))
-            }
-          }
-        }
-    }
-  }
-
-  // Visible for testing
-  def connectionRateForIp(ip: InetAddress): Int = {
-    connectionRatePerIp.getOrDefault(ip, defaultConnectionRatePerIp)
-  }
-
-  private[network] def addListener(config: KafkaConfig, listenerName: ListenerName): Unit = {
-    counts.synchronized {
-      if (!maxConnectionsPerListener.contains(listenerName)) {
-        val newListenerQuota = new ListenerConnectionQuota(counts, listenerName)
-        maxConnectionsPerListener.put(listenerName, newListenerQuota)
-        listenerCounts.put(listenerName, 0)
-        config.addReconfigurable(newListenerQuota)
-        newListenerQuota.configure(config.valuesWithPrefixOverride(listenerName.configPrefix))
-      }
-      counts.notifyAll()
-    }
-  }
-
-  private[network] def removeListener(config: KafkaConfig, listenerName: ListenerName): Unit = {
-    counts.synchronized {
-      maxConnectionsPerListener.remove(listenerName).foreach { listenerQuota =>
-        listenerCounts.remove(listenerName)
-        // once listener is removed from maxConnectionsPerListener, no metrics will be recorded into listener's sensor
-        // so it is safe to remove sensor here
-        listenerQuota.close()
-        counts.notifyAll() // wake up any waiting acceptors to close cleanly
-        config.removeReconfigurable(listenerQuota)
-      }
-    }
-  }
-
-  def dec(listenerName: ListenerName, address: InetAddress): Unit = {
-    counts.synchronized {
-      val count = counts.getOrElse(address,
-        throw new IllegalArgumentException(s"Attempted to decrease connection count for address with no connections, address: $address"))
-      if (count == 1)
-        counts.remove(address)
-      else
-        counts.put(address, count - 1)
-
-      if (totalCount <= 0)
-        error(s"Attempted to decrease total connection count for broker with no connections")
-      totalCount -= 1
-
-      if (maxConnectionsPerListener.contains(listenerName)) {
-        val listenerCount = listenerCounts(listenerName)
-        if (listenerCount == 0)
-          error(s"Attempted to decrease connection count for listener $listenerName with no connections")
-        else
-          listenerCounts.put(listenerName, listenerCount - 1)
-      }
-      counts.notifyAll() // wake up any acceptors waiting to process a new connection since listener connection limit was reached
-    }
-  }
-
-  def get(address: InetAddress): Int = counts.synchronized {
-    counts.getOrElse(address, 0)
-  }
-
-  private def waitForConnectionSlot(listenerName: ListenerName,
-                                    acceptorBlockedPercentMeter: com.yammer.metrics.core.Meter): Unit = {
-    counts.synchronized {
-      val startThrottleTimeMs = time.milliseconds
-      val throttleTimeMs = math.max(recordConnectionAndGetThrottleTimeMs(listenerName, startThrottleTimeMs), 0)
-
-      if (throttleTimeMs > 0 || !connectionSlotAvailable(listenerName)) {
-        val startNs = time.nanoseconds
-        val endThrottleTimeMs = startThrottleTimeMs + throttleTimeMs
-        var remainingThrottleTimeMs = throttleTimeMs
-        do {
-          counts.wait(remainingThrottleTimeMs)
-          remainingThrottleTimeMs = math.max(endThrottleTimeMs - time.milliseconds, 0)
-        } while (remainingThrottleTimeMs > 0 || !connectionSlotAvailable(listenerName))
-        acceptorBlockedPercentMeter.mark(time.nanoseconds - startNs)
-      }
-    }
-  }
-
-  // This is invoked in every poll iteration and we close one LRU connection in an iteration
-  // if necessary
-  def maxConnectionsExceeded(listenerName: ListenerName): Boolean = {
-    totalCount > brokerMaxConnections && !protectedListener(listenerName)
-  }
-
-  private def connectionSlotAvailable(listenerName: ListenerName): Boolean = {
-    if (listenerCounts(listenerName) >= maxListenerConnections(listenerName))
-      false
-    else if (protectedListener(listenerName))
-      true
-    else
-      totalCount < brokerMaxConnections
-  }
-
-  private def protectedListener(listenerName: ListenerName): Boolean =
-    interBrokerListenerName == listenerName && listenerCounts.size > 1
-
-  private def maxListenerConnections(listenerName: ListenerName): Int =
-    maxConnectionsPerListener.get(listenerName).map(_.maxConnections).getOrElse(Int.MaxValue)
-
-  /**
-   * Calculates the delay needed to bring the observed connection creation rate to listener-level limit or to broker-wide
-   * limit, whichever the longest. The delay is capped to the quota window size defined by QuotaWindowSizeSecondsProp
-   *
-   * @param listenerName listener for which calculate the delay
-   * @param timeMs current time in milliseconds
-   * @return delay in milliseconds
-   */
-  private def recordConnectionAndGetThrottleTimeMs(listenerName: ListenerName, timeMs: Long): Long = {
-    def recordAndGetListenerThrottleTime(minThrottleTimeMs: Int): Int = {
-      maxConnectionsPerListener
-        .get(listenerName)
-        .map { listenerQuota =>
-          val listenerThrottleTimeMs = recordAndGetThrottleTimeMs(listenerQuota.connectionRateSensor, timeMs)
-          val throttleTimeMs = math.max(minThrottleTimeMs, listenerThrottleTimeMs)
-          // record throttle time due to hitting connection rate quota
-          if (throttleTimeMs > 0) {
-            listenerQuota.listenerConnectionRateThrottleSensor.record(throttleTimeMs.toDouble, timeMs)
-          }
-          throttleTimeMs
-        }
-        .getOrElse(0)
-    }
-
-    if (protectedListener(listenerName)) {
-      recordAndGetListenerThrottleTime(0)
-    } else {
-      val brokerThrottleTimeMs = recordAndGetThrottleTimeMs(brokerConnectionRateSensor, timeMs)
-      recordAndGetListenerThrottleTime(brokerThrottleTimeMs)
-    }
-  }
-
-  /**
-   * Record IP throttle time on the corresponding listener. To avoid over-recording listener/broker connection rate, we
-   * also un-record the listener and broker connection if the IP gets throttled.
-   *
-   * @param listenerName listener to un-record connection
-   * @param throttleMs IP throttle time to record for listener
-   * @param timeMs current time in milliseconds
-   */
-  private def updateListenerMetrics(listenerName: ListenerName, throttleMs: Long, timeMs: Long): Unit = {
-    if (!protectedListener(listenerName)) {
-      brokerConnectionRateSensor.record(-1.0, timeMs, false)
-    }
-    maxConnectionsPerListener
-      .get(listenerName)
-      .foreach { listenerQuota =>
-        listenerQuota.ipConnectionRateThrottleSensor.record(throttleMs.toDouble, timeMs)
-        listenerQuota.connectionRateSensor.record(-1.0, timeMs, false)
-      }
-  }
-
-  /**
-   * Calculates the delay needed to bring the observed connection creation rate to the IP limit.
-   * If the connection would cause an IP quota violation, un-record the connection for both IP,
-   * listener, and broker connection rate and throw a ConnectionThrottledException. Calls to
-   * this function must be performed with the counts lock to ensure that reading the IP
-   * connection rate quota and creating the sensor's metric config is atomic.
-   *
-   * @param listenerName listener to unrecord connection if throttled
-   * @param address ip address to record connection
-   */
-  private def recordIpConnectionMaybeThrottle(listenerName: ListenerName, address: InetAddress): Unit = {
-    val connectionRateQuota = connectionRateForIp(address)
-    val quotaEnabled = connectionRateQuota != QuotaConfigs.IP_CONNECTION_RATE_DEFAULT
-    if (quotaEnabled) {
-      val sensor = getOrCreateConnectionRateQuotaSensor(connectionRateQuota, IpQuotaEntity(address))
-      val timeMs = time.milliseconds
-      val throttleMs = recordAndGetThrottleTimeMs(sensor, timeMs)
-      if (throttleMs > 0) {
-        trace(s"Throttling $address for $throttleMs ms")
-        // unrecord the connection since we won't accept the connection
-        sensor.record(-1.0, timeMs, false)
-        updateListenerMetrics(listenerName, throttleMs, timeMs)
-        throw new ConnectionThrottledException(address, timeMs, throttleMs)
-      }
-    }
-  }
-
-  /**
-   * Records a new connection into a given connection acceptance rate sensor 'sensor' and returns throttle time
-   * in milliseconds if quota got violated
-   * @param sensor sensor to record connection
-   * @param timeMs current time in milliseconds
-   * @return throttle time in milliseconds if quota got violated, otherwise 0
-   */
-  private def recordAndGetThrottleTimeMs(sensor: Sensor, timeMs: Long): Int = {
-    try {
-      sensor.record(1.0, timeMs)
-      0
-    } catch {
-      case e: QuotaViolationException =>
-        val throttleTimeMs = QuotaUtils.boundedThrottleTime(e, maxThrottleTimeMs, timeMs).toInt
-        debug(s"Quota violated for sensor (${sensor.name}). Delay time: $throttleTimeMs ms")
-        throttleTimeMs
-    }
-  }
-
-  /**
-   * Creates sensor for tracking the connection creation rate and corresponding connection rate quota for a given
-   * listener or broker-wide, if listener is not provided.
-   * @param quotaLimit connection creation rate quota
-   * @param connectionQuotaEntity entity to create the sensor for
-   */
-  private def getOrCreateConnectionRateQuotaSensor(quotaLimit: Int, connectionQuotaEntity: ConnectionQuotaEntity): Sensor = {
-    Option(metrics.getSensor(connectionQuotaEntity.sensorName)).getOrElse {
-      val sensor = metrics.sensor(
-        connectionQuotaEntity.sensorName,
-        rateQuotaMetricConfig(quotaLimit),
-        connectionQuotaEntity.sensorExpiration
-      )
-      sensor.add(connectionRateMetricName(connectionQuotaEntity), new Rate, null)
-      sensor
-    }
-  }
-
-  /**
-   * Updates quota configuration for a given connection quota entity
-   */
-  private def updateConnectionRateQuota(quotaLimit: Int, connectionQuotaEntity: ConnectionQuotaEntity): Unit = {
-    Option(metrics.metric(connectionRateMetricName(connectionQuotaEntity))).foreach { metric =>
-      metric.config(rateQuotaMetricConfig(quotaLimit))
-      info(s"Updated ${connectionQuotaEntity.metricName} max connection creation rate to $quotaLimit")
-    }
-  }
-
-  private def connectionRateMetricName(connectionQuotaEntity: ConnectionQuotaEntity): MetricName = {
-    metrics.metricName(
-      connectionQuotaEntity.metricName,
-      MetricsGroup,
-      s"Tracking rate of accepting new connections (per second)",
-      connectionQuotaEntity.metricTags.asJava)
-  }
-
-  private def rateQuotaMetricConfig(quotaLimit: Int): MetricConfig = {
-    new MetricConfig()
-      .timeWindow(config.quotaWindowSizeSeconds.toLong, TimeUnit.SECONDS)
-      .samples(config.numQuotaSamples)
-      .quota(new Quota(quotaLimit, true))
-  }
-
-  def close(): Unit = {
-    metrics.removeSensor(brokerConnectionRateSensor.name)
-    maxConnectionsPerListener.values.foreach(_.close())
-  }
-
-  class ListenerConnectionQuota(lock: Object, listener: ListenerName) extends ListenerReconfigurable with AutoCloseable {
-    @volatile private var _maxConnections = Int.MaxValue
-    private[network] val connectionRateSensor = getOrCreateConnectionRateQuotaSensor(Int.MaxValue, ListenerQuotaEntity(listener.value))
-    private[network] val listenerConnectionRateThrottleSensor = createConnectionRateThrottleSensor(ListenerThrottlePrefix)
-    private[network] val ipConnectionRateThrottleSensor = createConnectionRateThrottleSensor(IpThrottlePrefix)
-
-    def maxConnections: Int = _maxConnections
-
-    override def listenerName(): ListenerName = listener
-
-    override def configure(configs: util.Map[String, _]): Unit = {
-      _maxConnections = maxConnections(configs)
-      updateConnectionRateQuota(maxConnectionCreationRate(configs), ListenerQuotaEntity(listener.value))
-    }
-
-    override def reconfigurableConfigs(): util.Set[String] = {
-      SocketServer.ListenerReconfigurableConfigs.asJava
-    }
-
-    override def validateReconfiguration(configs: util.Map[String, _]): Unit = {
-      val value = maxConnections(configs)
-      if (value <= 0)
-        throw new ConfigException(s"Invalid ${KafkaConfig.MaxConnectionsProp} $value")
-
-      val rate = maxConnectionCreationRate(configs)
-      if (rate <= 0)
-        throw new ConfigException(s"Invalid ${KafkaConfig.MaxConnectionCreationRateProp} $rate")
-    }
-
-    override def reconfigure(configs: util.Map[String, _]): Unit = {
-      lock.synchronized {
-        _maxConnections = maxConnections(configs)
-        updateConnectionRateQuota(maxConnectionCreationRate(configs), ListenerQuotaEntity(listener.value))
-        lock.notifyAll()
-      }
-    }
-
-    def close(): Unit = {
-      metrics.removeSensor(connectionRateSensor.name)
-      metrics.removeSensor(listenerConnectionRateThrottleSensor.name)
-      metrics.removeSensor(ipConnectionRateThrottleSensor.name)
-    }
-
-    private def maxConnections(configs: util.Map[String, _]): Int = {
-      Option(configs.get(KafkaConfig.MaxConnectionsProp)).map(_.toString.toInt).getOrElse(Int.MaxValue)
-    }
-
-    private def maxConnectionCreationRate(configs: util.Map[String, _]): Int = {
-      Option(configs.get(KafkaConfig.MaxConnectionCreationRateProp)).map(_.toString.toInt).getOrElse(Int.MaxValue)
-    }
-
-    /**
-     * Creates sensor for tracking the average throttle time on this listener due to hitting broker/listener connection
-     * rate or IP connection rate quota. The average is out of all throttle times > 0, which is consistent with the
-     * bandwidth and request quota throttle time metrics.
-     */
-    private def createConnectionRateThrottleSensor(throttlePrefix: String): Sensor = {
-      val sensor = metrics.sensor(s"${throttlePrefix}ConnectionRateThrottleTime-${listener.value}")
-      val metricName = metrics.metricName(s"${throttlePrefix}connection-accept-throttle-time",
-        MetricsGroup,
-        "Tracking average throttle-time, out of non-zero throttle times, per listener",
-        Map(ListenerMetricTag -> listener.value).asJava)
-      sensor.add(metricName, new Avg)
-      sensor
-    }
-  }
-}
-
-class TooManyConnectionsException(val ip: InetAddress, val count: Int) extends KafkaException(s"Too many connections from $ip (maximum = $count)")
-
-class ConnectionThrottledException(val ip: InetAddress, val startThrottleTimeMs: Long, val throttleTimeMs: Long)
-  extends KafkaException(s"$ip throttled for $throttleTimeMs")
