+++ core/src/main/scala/kafka/network/TooManyConnectionsException.scala
@@ -1,0 +1,7 @@
+package kafka.network
+
+import org.apache.kafka.common.KafkaException
+
+import java.net.InetAddress
+
+class TooManyConnectionsException(val ip: InetAddress, val count: Int) extends KafkaException(s"Too many connections from $ip (maximum = $count)")
