+++ core/src/main/scala/kafka/server/AutoTopicCreationManager.scala
@@ -255,11 +255,13 @@ class DefaultAutoTopicCreationManager(
 
   private def convertToTopicConfigCollections(config: Properties): CreateableTopicConfigCollection = {
     val topicConfigs = new CreateableTopicConfigCollection()
-    config.forEach {
-      case (name, value) =>
+    config.asScala.foreach {
+      (name: Object, value: Object) => {
         topicConfigs.add(new CreateableTopicConfig()
           .setName(name.toString)
           .setValue(value.toString))
+        Unit
+      }
     }
     topicConfigs
   }
