+++ core/src/main/scala/kafka/server/ReplicaManager.scala
@@ -1926,7 +1926,7 @@ class ReplicaManager(val config: KafkaConfig,
   protected def createReplicaSelector(): Option[ReplicaSelector] = {
     config.replicaSelectorClassName.map { className =>
       val tmpReplicaSelector: ReplicaSelector = CoreUtils.createObject[ReplicaSelector](className)
-      tmpReplicaSelector.configure(config.originals())
+      tmpReplicaSelector.configure(config.originals)
       tmpReplicaSelector
     }
   }
